+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-####17.3 (18 *alpha*) - NOT RELEASED
+####18 - NOT RELEASED
+* Added sharing! Share all your passwords with others (you can trust)!
+ * The users you can share with, is based on the admin settings (only from your own group, or all users, ...)
+ * Icons indicate to how many users you've shared a specific password
+ * It uses a random share key (32 hexadecimal characters) that is created everytime a share is created. This key is saved to a new (third) database table, `oc_passwords_share`, and to the encrypted `properties` column of the password owner. When the keys match, the password will be decrypted on the other user's side.
 * This app can now fully be controlled remotely! This makes it technically possible to use ownCloud Passwords on Android, iPhones, remote servers, you name it. Other authors have already made browser plugins available for Firefox and Chrome. No strict need to use the website of ownCloud anymore, but it all works just as safe. 
  * Changed RESTful API to support GET, POST, DELETE, PUT, PATCH
  * Moved all calculation classes to server-side (translated JavaScript to PHP, which is all PHP 7 safe)
@@ -6,12 +10,16 @@
  * Firefox addon: [here](https://addons.mozilla.org/en-US/firefox/addon/firefox-owncloud-passwords) (thanks to [@eglia](https://github.com/eglia)) 
  * Chrome extension: [here](https://github.com/thefirstofthe300/ownCloud-Passwords) (thanks to [@thefirstofthe300](https://github.com/thefirstofthe300))
 * Allow tabs for input in notes field (so pressing Tab doesn't switch to another field, but instead inserts a tab)
-* Filtering a category now only searches active passwords (ignores passwords in trash bin)
+* Filtering a category or text now only searches active passwords, ignoring passwords in the trash bin
+* Added 'Edit categories' button to category popup
 * Changed all deprecated PHP classes, to follow ownCloud's guidelines
+* Dropped support for PostgreSQL, for now (I'll try to support it from 18.1 on again)
 * Fix for column headers `Strength` and `Last changed`
 * Fix for scrollbar on sidebar
 * Fix for reset of category list after adding a password
 * Fix for losing a full URL when password was changed
+* Fix for popup on smaller screens (mobile phones): the popup is now scrollable when it covers more than 75% of the browsers height
+* Changed popup background to better show buttons like 'Generate password'
 
 ####17.2 - Mar 12, 2016
 * Fix for saving and updating a password on PostgreSQL backends